+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog #
 
+## Version 2.6.3 ##
+
+🛠️ Bug fixes:
+
+* Fixed typing issues in `plotpy.events` module:
+  * Event objects were not properly typed (`QtCore.QEvent` instead of `QtGui.QMouseEvent`)
+  * Event position arguments were not properly typed (`QtCore.QPoint` instead of `QtCore.QPointF`)
+
 ## Version 2.6.2 ##
 
 💥 New features / Enhancements: